Moved the api prefix to the endpoints to accomodate Lemmy Pictrs API
Lemmy does not expose Pictrs directly, but instead it proxies the requests.
Ideally Pictrs could be a separate OpenAPI spec, but Lemmy doesn't expose raw
Pictrs API - despite the endpoint name (`/pictrs/image`) implying it. Instead
Lemmy proxies the requests, so this is essentially Lemmy own api for managing
images. In that case the api is misplaced, unfortunately. The `/pictrs/image`
should really be `/api/v3/image`, but there is nothing we can do right now.